+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # **Upcoming release**
 
+- #710, #561 Implement `except*` syntax
+
 # Release 1.10.0
 
 - #708, #709 Add support for Python 3.12