+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # **Upcoming release**
 
-- #710, #561 Implement `except*` syntax
+- #710, #561 Implement `except*` syntax (@lieryan)
+- #711 allow building documentation without having rope module installed (@kloczek)
 
 # Release 1.10.0
 
-- #708, #709 Add support for Python 3.12
+- #708, #709 Add support for Python 3.12 (@lieryan)
 
 # Release 1.9.0