+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - #710, #561 Implement `except*` syntax (@lieryan)
 - #711 allow building documentation without having rope module installed (@kloczek)
-- #719 Create in-memory databases with shared cache per project to support a one-thread-one-autoimport model in multithreading environments (@tkrabel)
+- #719 Allow in-memory databases being shared across threads (@tkrabel)
 
 # Release 1.10.0