+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - #710, #561 Implement `except*` syntax (@lieryan)
 - #711 allow building documentation without having rope module installed (@kloczek)
+- #722 speed up module cache generation by factor of 2 (@tkrabel)
 
 # Release 1.10.0